% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add Two Numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{one number}

\item{y}{one number \code{\link[base]{sum}}}
}
\value{
sum of two numbers
}
\description{
XXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
YYYYYYYYYYYYYYYYYYYYYYYYYYYYY
ZZZZZZZZZZZZZZZZZZZZZZZZ.
}
\examples{
sum(1, 2)
\donttest{
sum(3, 4)}


}
